misc(generators): refactor
